# Remove a bounds check in pure Go